New Transaction Tables successfully migrated
