+++ usaspending_api/awards/management/commands/update_transaction_fiscal_year.py
@@ -1,4 +1,4 @@
-from usaspending_api.broker.models import TransactionNormalized
+from usaspending_api.awards.models import TransactionNormalized
 from django.core.management.base import BaseCommand
 import logging
