+++ usaspending_api/awards/tests/test_transactions.py
@@ -5,7 +5,7 @@ from model_mommy import mommy
 import pytest
 from rest_framework import status
 
-from usaspending_api.broker.models import TransactionNormalized, TransactionFABS, TransactionFPDS
+from usaspending_api.awards.models import TransactionNormalized, TransactionFABS, TransactionFPDS
 
 
 @pytest.mark.django_db
