+++ usaspending_api/awards/v1/views.py
@@ -1,7 +1,7 @@
 from collections import namedtuple
 
 from usaspending_api.awards.models import Award, Subaward
-from usaspending_api.broker.models import TransactionNormalized
+from usaspending_api.awards.models import TransactionNormalized
 from usaspending_api.awards.serializers import AwardSerializer, SubawardSerializer, TransactionNormalizedSerializer
 from usaspending_api.common.mixins import FilterQuerysetMixin, SuperLoggingMixin, AggregateQuerysetMixin
 from usaspending_api.common.serializers import AggregateSerializer
