+++ usaspending_api/etl/management/commands/load_submission.py
@@ -15,7 +15,7 @@ from usaspending_api.accounts.models import (
     AppropriationAccountBalances, AppropriationAccountBalancesQuarterly,
     TreasuryAppropriationAccount)
 from usaspending_api.awards.models import Award, FinancialAccountsByAwards
-from usaspending_api.broker.models import TransactionNormalized, TransactionFABS
+from usaspending_api.awards.models import TransactionNormalized, TransactionFABS
 from usaspending_api.financial_activities.models import (
     FinancialAccountsByProgramActivityObjectClass, TasProgramActivityObjectClassQuarterly)
 from usaspending_api.references.models import (
