+++ usaspending_api/etl/management/load_base.py
@@ -18,7 +18,7 @@ from django.db import connection, connections, utils
 from django.core.cache import caches
 
 from usaspending_api.awards.models import Award
-from usaspending_api.broker.models import TransactionNormalized, TransactionFABS, TransactionFPDS
+from usaspending_api.awards.models import TransactionNormalized, TransactionFABS, TransactionFPDS
 from usaspending_api.etl.award_helpers import (
     update_awards, update_contract_awards,
     update_award_categories, )
