+++ usaspending_api/etl/tests/test_subaward_etl.py
@@ -6,7 +6,7 @@ from django.core.management import call_command
 from model_mommy import mommy
 
 from usaspending_api.awards.models import Award, Subaward
-from usaspending_api.broker.models import TransactionNormalized, TransactionFABS, TransactionFPDS
+from usaspending_api.awards.models import TransactionNormalized, TransactionFABS, TransactionFPDS
 from usaspending_api.submissions.models import SubmissionAttributes
