+++ usaspending_api/references/management/commands/update_location_usage_flags.py
@@ -5,7 +5,7 @@ from django.db.models import Q
 
 from usaspending_api.references.models import Location, LegalEntity
 from usaspending_api.awards.models import Award
-from usaspending_api.broker.models import TransactionNormalized
+from usaspending_api.awards.models import TransactionNormalized
 
 
 class Command(BaseCommand):
