+++ usaspending_api/references/tests/test_psc_autocomplete_v2.py
@@ -4,7 +4,7 @@ import pytest
 from model_mommy import mommy
 from rest_framework import status
 
-from usaspending_api.broker.models import TransactionFPDS
+from usaspending_api.awards.models import TransactionFPDS
 
 
 @pytest.fixture
