+++ usaspending_api/references/v2/views/autocomplete.py
@@ -4,8 +4,7 @@ from django.db.models.functions import Greatest
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from usaspending_api.awards.models import LegalEntity, TreasuryAppropriationAccount
-from usaspending_api.broker.models import TransactionFPDS
+from usaspending_api.awards.models import LegalEntity, TreasuryAppropriationAccount, TransactionFPDS
 from usaspending_api.common.exceptions import InvalidParameterException
 from usaspending_api.references.models import Agency, Cfda
 from usaspending_api.references.v1.serializers import AgencySerializer
