+++ usaspending_api/search/tests/test_spending_by_category.py
@@ -4,8 +4,7 @@ import pytest
 from model_mommy import mommy
 from rest_framework import status
 
-from usaspending_api.awards.models import Award
-from usaspending_api.broker.models import TransactionNormalized
+from usaspending_api.awards.models import Award, TransactionNormalized
 from usaspending_api.references.models import Location, Agency, ToptierAgency, SubtierAgency
