+++ usaspending_api/submissions/management/commands/submission_stats.py
@@ -5,8 +5,8 @@ from django.core.exceptions import ObjectDoesNotExist
 
 from usaspending_api.accounts.models import (
     AppropriationAccountBalances, AppropriationAccountBalancesQuarterly)
-from usaspending_api.awards.models import Award, FinancialAccountsByAwards
-from usaspending_api.broker.models import TransactionNormalized, TransactionFPDS, TransactionFABS
+from usaspending_api.awards.models import (Award, FinancialAccountsByAwards,
+                                           TransactionNormalized, TransactionFPDS, TransactionFABS)
 from usaspending_api.financial_activities.models import (
     FinancialAccountsByProgramActivityObjectClass, TasProgramActivityObjectClassQuarterly
 )
